_search-input: sass: avoid using slash as div
Sass currently treats / as a division operation in some contexts and a
separator in others. This makes it difficult for Sass users to tell what
any given / will mean, and makes it hard to work with new CSS features
that use / as a separator.

As such, Sass has deprecated its usage, and `dart-sass` (which we use in
this project) flags it at build time.

To fix this issue, let's prefer `math.div` instead.

See: https://sass-lang.com/documentation/breaking-changes/slash-div for
more details.

This fixes issue #1276.