package: remove `ember-collapsible-panel` dep
At build-time, a warning is logged:
```
Usages of the Ember Global may be caused by an outdated ember-cli-babel
dependency. The following steps may help:
* Upgrade the following addons to the latest version:
  ...
  * ember-collapsible-panel
```

We *do* depend on the `ember-collapsible-panel` package, but do not
appear to actually use it.  For reference, it looks like it was added in
`Commit: aecc6b5d`. and removed in `Commit: dda4d0e5`.

As such, we should just be able to remove the dependency instead of
trying to upgrade it.

This should help along the way to achieving:
https://github.com/ember-learn/deprecation-app/issues/1272.