package: remove ember/render-modifiers devDep
Right now, we depend on the @ember/render-modifiers package, but do not
actually use it.  For reference, it looks like the original dep/usage
was added in `Commit: c0f373b4` and then that usage was subsequently
removed in `Commit: 459a5769`.

As such, we should just be able to remove the dependency.