config/dependency-lint: remove `ember-concurrency` version
For context, `ember/dependency-lint` is an addon that adds lint tests
that verify only one version of any given addon will be activated in the
final built application.

Right now, we've explicitly overriden that behavior for
`ember-concurrency` in the config/dependency-lint.js file to allow
either version 2.1.2 or version 1.3.0 to be installed. Given this
configuration, if any other version is installed, ember/dependency-lint
will throw a linter error.

After updating `ember-styleguide` we only depend on v2.3.7 of the
ember-concurrency package, so no extra configuration should be required
for this package to lint successfully. As such, let's go ahead and
remove this setting in the config file.